requirements.txt
matplotlib==3.3.2
streamlit~=1.28.2
numpy~=1.26.2
Keras~=2.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
Pillow~=10.1.0